requirements.txt
pandas
sqlalchemy
pymysql
streamlit
python-dotenv
plotly
nba_api